Use bitvectors instead of boolean vectors
Before, the Pauli frames were stored as an array of booleans (for
clarity). I switched to bitvectors. This should reduce the storage for
the frames, by a factor of 8. The tracking is probably also a little bit
faster, but I guess that wasn't really a problem before. The
disadvantage is, that we need additional logic to read the frames (cf.
bitvec_to_vec_bool function in the check.py script), but it is not that
difficult.
I also removed the transposed frames, because we are not using it.